Remove nulls (#53)
