+++ ingestor-default/src/test/scala/za/co/absa/hyperdrive/ingestor/implementation/reader/kafka/TestKafkaStreamReader.scala
@@ -38,26 +38,15 @@ class TestKafkaStreamReader extends FlatSpec with MockitoSugar {
   behavior of "KafkaStreamReader"
 
   it should "throw on blank topic" in {
-    assertThrows[IllegalArgumentException]( // null topic
-      new KafkaStreamReader(topic = null, validBrokers, validExtraConfs)
-    )
     assertThrows[IllegalArgumentException]( // empty topic
       new KafkaStreamReader(topic = "  ", validBrokers, validExtraConfs)
     )
   }
 
   it should "throw on blank brokers" in {
-    assertThrows[IllegalArgumentException]( // null topic
-      new KafkaStreamReader(validTopic, brokers = null, validExtraConfs)
-    )
     assertThrows[IllegalArgumentException]( // empty topic
       new KafkaStreamReader(validTopic, brokers = "  ", validExtraConfs)
     )
-  }
-
-  it should "throw on null SparkSession" in {
-    val reader = new KafkaStreamReader(validTopic, validBrokers, validExtraConfs)
-    assertThrows[IllegalArgumentException](reader.read(spark = null))
   }
 
   it should "throw if SparkSession is stopped" in {
@@ -97,7 +86,7 @@ class TestKafkaStreamReader extends FlatSpec with MockitoSugar {
     verify(dataStreamReader).option(SPARK_BROKERS_SETTING_KEY, validBrokers)
     verify(dataStreamReader, never()).options(validExtraConfs)
   }
-  
+
   private def getMockedSparkContext(stopped: Boolean): SparkContext = {
     val sparkContext = mock[SparkContext]
     when(sparkContext.isStopped).thenReturn(stopped)
